# backend/jvm/webapi/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Ticket Management API
  version: 1.0.0
servers:
  - url: http://localhost:8080/
...
paths:
...
  /categories:
    get:
      tags:
        - Categories
      description: "Get all Categories"
      operationId: "getAllCategories"
      security:
        - basicAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags:
        - Categories
      description: "Add a new Category"
      operationId: "createCategory"
      security:
        - basicAuth: []
      requestBody:
        description: Category object to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryPost'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/{categoryId}:
    get:
      tags:
        - Categories
      description: "Get Category by Id"
      operationId: "getCategoryById"
      security:
        - basicAuth: []
      parameters:
        - $ref: "#/components/parameters/category_id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    put:
      tags:
        - Categories
      description: "Update Category by Id"
      operationId: "updateCategoryById"
      security:
        - basicAuth: []
      parameters:
        - $ref: "#/components/parameters/category_id"
      requestBody:
        description: Category object to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    delete:
      tags:
        - Categories
      description: "Delete Category by Id"
      operationId: "deleteCategoryById"
      security:
        - basicAuth: []
      parameters:
        - $ref: "#/components/parameters/category_id"
      responses:
        "204":
          description: No Content
...
  /signup:
    post:
      tags:
        - Authentication
      summary: Create a new user
      operationId: createUser
      requestBody:
        description: User object to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPost'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  parameters:
...
    category_id:
      name: categoryId
      in: path
      description: ID of category to retrieve
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    Category:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        eta:
          type: integer
          description: "Estimated Time of Accomplishment"
        priority:
          type: string
          enum:
            - URGENT
            - HIGH
            - MEDIUM
            - LOW
    CategoryPost:
      type: object
      properties:
        name:
          type: string
        eta:
          type: integer
          description: "Estimated Time of Accomplishment"
        priority:
          type: string
          enum:
            - URGENT
            - HIGH
            - MEDIUM
            - LOW
...
    User:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          readOnly: true
        username:
          type: string
        fname:
          type: string
        lname:
          type: string
        email:
          type: string
        password:
          type: string
    UserPost:
      type: object
      properties:
        username:
          type: string
        fname:
          type: string
        lname:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - USER
            - EMPLOYEE
            - ADMIN
          default: USER
...